Working on Auth

+++ app/Http/Controllers/PostController.php
@@ -3,10 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Post;
+use Auth;
 use App\Like;
 use App\Tag;
 use Illuminate\Http\Request;
-
 use App\Http\Requests;
 
 class PostController extends Controller
@@ -19,19 +19,22 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
 
     public function getPost($id)
     {
-        $post = Post::where('id','=',$id)->with->first();
+        $post = Post::where('id', '=', $id)->with('likes')->first();
         return view('blog.post', ['post' => $post]);
     }
 
     public function getLikePost($id)
     {
-        $post = Post::where('id',$id)->first();
+        $post = Post::where('id', $id)->first();
         $like = new Like();
         $post->likes()->save($like);
         return redirect()->back();
@@ -39,37 +42,50 @@ class PostController extends Controller
 
     public function getAdminCreate()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
-        return view('admin.create',['tags' => $tags]);
+        return view('admin.create', ['tags' => $tags]);
     }
 
-    public function getAdminEdit( $id)
+    public function getAdminEdit($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
     }
 
-    public function postAdminCreate( Request $request)
+    public function postAdminCreate(Request $request)
     {
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
+        if (!$user) {
+            return redirect()->back();
+        }
         $post = new Post([
-            'title' =>  $request->input('title'), 
+            'title' => $request->input('title'),
             'content' => $request->input('content')
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
-        
+
         return redirect()->route('admin.index')
-        ->with('info','Post created, Title is:'
-        .$request->input('title'));
+            ->with('info', 'Post created, Title is:'
+                . $request->input('title'));
     }
 
-    public function postAdminUpdate( Request $request)
+    public function postAdminUpdate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
@@ -80,17 +96,20 @@ class PostController extends Controller
         $post->save();
         $post->tags()->sync($request->input('tags') === null ? [] : $request->input('tags'));
         return redirect()->route('admin.index')
-        ->with('info','Post edited, new Title is:'.$request
-        ->input('title'));
+            ->with('info', 'Post edited, new Title is:' . $request
+                ->input('title'));
     }
 
-    public function getAdminDelete ($id)
+    public function getAdminDelete($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
         return redirect()->route('admin.index')
-        ->with('info','Post deleted');
+            ->with('info', 'Post deleted');
     }
 }
